Refactor: split capture and reader
